lsm: refactor compaction
Refactor compaction to:
- make the code easier to follow,
- make it less unobvious how the compactions are actually schedules
  (that is, what work can be in-flight at any given point in time),
- make scheduling a bit more aggressive inside a single compaction (e.g.
  start merge when the first two value blocks are in, separate IO and
  block resources),
- make it easier to run several compactions at the same time.

Code Changes
============

* Fewer types
* All tricky logic is centralized in `compaction_dispatch`
* Many FIFOs are replaced with fewer RingBuffers
* Flatten Beat and Bar into compaction

CompactionPipeline no longer implements pipelining. Now, it only deals
with scheduling compactions themselves. Pipelining is the job of each
individual compaction. That is, CompactonPipeline kicks off a
particualar compaciton, and recieves the control back only when that
compaciton's beat is done.

CompactionPipeline should be rename to CompactionSchedule, or maybe just
merged with a forest.

A new type, CompactionPool, is introduced. This is a bag of all resources
shared by compactions. Previously, blocks, reads and writes were lumped
together. Now, CompactionPool holds separate IOPS for reads, writes,
blocks, and CPUS.

CompactionPool is a forest-level object, but it _doesn't_ depend on the
type of the forest. So, all compactions share the same CompactionPool
_type_, which allows them to share the actual object.

The actual pipelining logic is moved to `Compaction.compaction_dispatch`
function. This function no longer tries to _explicitly_ encode a
pipeline. Rather, it is just a loop that spawns a bunch of jobs, and the
pipelining happens as a side-effect of that. The pseudo-code for
compaction_dispatch is:

    for each job:
        if have resources for the job AND
           job's dependencies are fulfilled AND
           it makes sense to start the job:

           start the job

Here, the jobs are reading the index blocks, reading the value blocks,
merging/copying, writing the data blocks, writing the index blocks.
Resources are blocks (memory) and grid reads&writes. Dependencies are
blocks read previously by other jobs (reading a value block needs an
index block). The "makes sense" part coves:
- spreading resources across jobs to avoid deadlocks,
- avoiding overreads when finishing a beat-worth of work.

There _isn't_ some kind of comptime meta programming framework for
defining jobs, it's just a good old pile of ifs (pushed up to the
`compaction_dispatch` cental location).

Dependency tracking is done via RingBuffers. For example, there's a
RingBuffer of level_b value blocks. Blocks are added to this queue in
order when we _start_ the corrsponding read. When we finish the read, we
flip the status of block (which is in the queue at this point) to
`read_done`.

So, when we think about launching a merge, we check that:

- there's a head value block in the queue
- the status of this block is `read_done`

If the merge ends up fully consuming the block, it pops it off the queue
and adds back to the shared pool.

This setup is similar to a bunch of fifos we used before, but:

* With ring-buffers, so that it's statically obvious what's the upper
  bound for the size of the queue.
* Queues are matched to block kind (value block/index block/output block)
  rather than to it's lifetime stage (pending/reading etc).
* There are just fewer queues (I think, didn't actually count).

Behavior Changes
================

* Separate limits for memory, CPU, and IO resources.
* Aggressive scheduling -- start merge as soon as the first two blocks
  are in
* Running two compactions just work, given enough resources.
* A different "winding down" approach.

Previously, allocating a block for compaction also allocated a grid read
and a grid write. In the new approach, those are allocated separately.
Additionally, besides available resources there are static limits on the
amount of outstanding work. That is, RingBuffers for blocks are
statically sized, so you can't read more than X blocks at the same time
even if you have spare blocks.

For simplicity, it is assumed that there's actually enough reads and
writes to fill the queues fully, but the approach should work even if
you restrict _all_ compactions to just a single read and a single write.

The blocks are distributed in a priority round-robin fashion. That is,
we first give block to the table builder, and then alternate between
level b and level a. This way:

* We avoid deadlocks: if the number of blocks is larger than a minimum,
  than the jobs will be scheduled in the right order to avoid running
  out of blocks completely
* We give equal resources to level a and level b (the approach should be
  extensible to fair allocation, proportional to level "size")
* The schedule ends up being maximally pipelined (the first four reads
  are: level_a index, level_b index, level_b value, level_a value,
  which allows starting the merge ASAP).

I _think_ this schedule allows even for a simplistic strategy for
running several compactions: if you have N*min blocks, it would be
correct to start N compactions at the same time. At least one of them
will be guaranteed to be able to proceed. This scheme is simplistic,
because:
- a single compaction can monopolize an unfair share of resources.
  Compactions are greedy.
- there's no code path for "if compaction A frees resource X, wake up
  compaction B waiting on X".

Both points should be fixable I think though!

Another difference (I think) happens in the "winding down" phase. That
is, in what happens when the compaction consumed the required quota of
inputs, but haven't yet produced a full value table of output. This is a
tricky situation, because, depending on the tombstones, you might have
to read quite a bit more tables still. The worst case here is that you
have a full table on level A, a full table on level B, but their merge
is an empty set.

In the previous approach, due to explicit pipelining, we'll just loop
over:

    loop {
        read fixed number of blocks from level
        run merge
    }

In the new approach, this is handled by `AND it makes sense to start the
job` part of the pseudocode --- if we _are_ in the winding-down phase,
we just artificially limit the depth of our RingBuffers.

Such overreads are tracked via `values_wasted_count`.